refactor: implement cost calculation and input sanitization for manpower rates
